- Script zum Aufspüren von unbenutzten Bildern hinzugefügt. - Eine Reihe unbenutzter Bilder entfernt. - Neues Refresh Bild hinzugefügt. - Alle Warnungen gefixt. - Aufräumarbeiten in MainMenu + BankingController. - Geänderte Icons in den Preferences.

+++ client/Source/NSString+PecuniaAdditions.m
@@ -41,7 +41,7 @@
         
         NSMutableString *hexString  = [NSMutableString stringWithCapacity: 2 * data.length];
         for (NSUInteger i = 0; i < data.length; ++i)
-            [hexString appendString:[NSString stringWithFormat: @"%02x", (unsigned long)dataBuffer[i]]];
+            [hexString appendString:[NSString stringWithFormat: @"%02x", dataBuffer[i]]];
         result = [NSString stringWithString: hexString];
     }
     return result;
